[FEAT] 약속잡기 - 친구 검색 뷰 작업
