--- momogum/Appoint/AppointFriendSearchView.swift
@@ -1,24 +1,0 @@
-//
-//  AppointFriendSearchView.swift
-//  momogum
-//
-//  Created by nelime on 1/17/25.
-//
-
-import SwiftUI
-
-struct AppointFriendSearchView: View {
-    @State var searchText: String = ""
-    
-    var body: some View {
-        ScrollView{
-            VStack {
-                EmptyView()
-            }
-        }
-    }
-}
-
-#Preview {
-    AppointFriendSearchView()
-}
